Simplify `->` method tail handling, ensuring left associativity (backport #7144) (#7195)
