+++ apollo-federation/src/sources/connect/json_selection/methods.rs
@@ -5,7 +5,6 @@ use shape::location::SourceId;
 
 use super::ApplyToError;
 use super::MethodArgs;
-use super::PathList;
 use super::VarsWithPathsMap;
 use super::immutable::InputPath;
 use super::location::WithRange;
@@ -59,7 +58,7 @@ pub(super) enum ArrowMethod {
 macro_rules! impl_arrow_method {
     ($struct_name:ident, $impl_fn_name:ident, $shape_fn_name:ident) => {
         #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-        pub(super) struct $struct_name;
+        pub(crate) struct $struct_name;
         impl $crate::sources::connect::json_selection::methods::ArrowMethodImpl for $struct_name {
             fn apply(
                 &self,
@@ -68,9 +67,8 @@ macro_rules! impl_arrow_method {
                 data: &JSON,
                 vars: &VarsWithPathsMap,
                 input_path: &InputPath<JSON>,
-                tail: &WithRange<PathList>,
             ) -> (Option<JSON>, Vec<ApplyToError>) {
-                $impl_fn_name(method_name, method_args, data, vars, input_path, tail)
+                $impl_fn_name(method_name, method_args, data, vars, input_path)
             }
 
             fn shape(
@@ -104,7 +102,6 @@ pub(super) trait ArrowMethodImpl {
         data: &JSON,
         vars: &VarsWithPathsMap,
         input_path: &InputPath<JSON>,
-        tail: &WithRange<PathList>,
     ) -> (Option<JSON>, Vec<ApplyToError>);
 
     fn shape(
