+++ apollo-federation/src/sources/connect/json_selection/methods/public/mod.rs
@@ -1,0 +1,19 @@
+mod echo;
+pub(crate) use echo::EchoMethod;
+mod entries;
+pub(crate) use entries::EntriesMethod;
+mod first;
+pub(crate) use first::FirstMethod;
+mod json_stringify;
+pub(crate) use json_stringify::JsonStringifyMethod;
+mod last;
+pub(crate) use last::LastMethod;
+mod map;
+pub(crate) use map::MapMethod;
+mod r#match;
+pub(crate) use r#match::MatchMethod;
+pub(crate) use r#match::match_shape;
+mod size;
+pub(crate) use size::SizeMethod;
+mod slice;
+pub(crate) use slice::SliceMethod;
